refactor: isolate tasks and task comments as separate components (#976)
* Refactor: isolate tasks and task comments as the separate components and reformat response messages with f-strings 
* Small change in task controller description
