+++ app/api/resources/mentorship_relation.py
@@ -2,18 +2,12 @@ from flask import request
 from flask_restx import Resource, Namespace, marshal
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from http import HTTPStatus
+
 from app import messages
-from app.api.dao.task import TaskDAO
-from app.api.dao.task_comment import TaskCommentDAO
 from app.api.resources.common import auth_header_parser
 from app.api.dao.mentorship_relation import MentorshipRelationDAO
 from app.api.dao.user import UserDAO
 from app.api.models.mentorship_relation import *
-from app.api.validations.task_comment import (
-    validate_task_comment_request_data,
-    COMMENT_MAX_LENGTH,
-)
-from app.utils.validation_utils import get_length_validation_error_message
 from app.database.models.mentorship_relation import MentorshipRelationModel
 from app.api.email_utils import send_email_mentorship_relation_accepted
 from app.api.email_utils import send_email_new_request
@@ -35,40 +29,32 @@ class SendRequest(Resource):
     @mentorship_relation_ns.doc("send_request")
     @mentorship_relation_ns.expect(auth_header_parser, send_mentorship_request_body)
     @mentorship_relation_ns.response(
-        HTTPStatus.CREATED.value,
-        "%s" % messages.MENTORSHIP_RELATION_WAS_SENT_SUCCESSFULLY,
+        HTTPStatus.CREATED, f"{messages.MENTORSHIP_RELATION_WAS_SENT_SUCCESSFULLY}"
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.BAD_REQUEST.value,
-        "%s\n%s\n%s\n%s\n%s\n%s\n%s\n%s\n%s\n%s\n%s\n%s\n%s"
-        % (
-            messages.MATCH_EITHER_MENTOR_OR_MENTEE,
-            messages.MENTOR_ID_SAME_AS_MENTEE_ID,
-            messages.END_TIME_BEFORE_PRESENT,
-            messages.MENTOR_TIME_GREATER_THAN_MAX_TIME,
-            messages.MENTOR_TIME_LESS_THAN_MIN_TIME,
-            messages.MENTOR_NOT_AVAILABLE_TO_MENTOR,
-            messages.MENTEE_NOT_AVAIL_TO_BE_MENTORED,
-            messages.MENTOR_ALREADY_IN_A_RELATION,
-            messages.MENTEE_ALREADY_IN_A_RELATION,
-            messages.MENTOR_ID_FIELD_IS_MISSING,
-            messages.MENTEE_ID_FIELD_IS_MISSING,
-            messages.NOTES_FIELD_IS_MISSING,
-            messages.END_DATE_FIELD_IS_MISSING,
-        ),
+        HTTPStatus.BAD_REQUEST,
+        f"{messages.MATCH_EITHER_MENTOR_OR_MENTEE}\n"
+        f"{messages.MENTOR_ID_SAME_AS_MENTEE_ID}\n"
+        f"{messages.END_TIME_BEFORE_PRESENT}\n"
+        f"{messages.MENTOR_TIME_GREATER_THAN_MAX_TIME}\n"
+        f"{messages.MENTOR_TIME_LESS_THAN_MIN_TIME}\n"
+        f"{messages.MENTOR_NOT_AVAILABLE_TO_MENTOR}\n"
+        f"{messages.MENTEE_NOT_AVAIL_TO_BE_MENTORED}\n"
+        f"{messages.MENTOR_ALREADY_IN_A_RELATION}\n"
+        f"{messages.MENTEE_ALREADY_IN_A_RELATION}\n"
+        f"{messages.MENTOR_ID_FIELD_IS_MISSING}\n"
+        f"{messages.MENTEE_ID_FIELD_IS_MISSING}\n"
+        f"{messages.NOTES_FIELD_IS_MISSING}",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-        ),
+        HTTPStatus.UNAUTHORIZED,
+        f"{messages.TOKEN_HAS_EXPIRED}\n"
+        f"{messages.TOKEN_IS_INVALID}\n"
+        f"{messages.AUTHORISATION_TOKEN_IS_MISSING}",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND.value,
-        "%s\n%s" % (messages.MENTOR_DOES_NOT_EXIST, messages.MENTEE_DOES_NOT_EXIST),
+        HTTPStatus.NOT_FOUND,
+        f"{messages.MENTOR_DOES_NOT_EXIST}\n" f"{messages.MENTEE_DOES_NOT_EXIST}",
     )
     def post(cls):
         """
@@ -150,13 +136,10 @@ class GetAllMyMentorshipRelation(Resource):
         model=mentorship_request_response_body,
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-        ),
+        HTTPStatus.UNAUTHORIZED,
+        f"{messages.TOKEN_HAS_EXPIRED}\n"
+        f"{messages.TOKEN_IS_INVALID}\n"
+        f"{messages.AUTHORISATION_TOKEN_IS_MISSING}\n",
     )
     @mentorship_relation_ns.marshal_list_with(
         mentorship_request_response_body,
@@ -195,31 +178,23 @@ class AcceptMentorshipRelation(Resource):
     @mentorship_relation_ns.doc("accept_mentorship_relation")
     @mentorship_relation_ns.expect(auth_header_parser)
     @mentorship_relation_ns.response(
-        HTTPStatus.OK.value,
-        "%s" % messages.MENTORSHIP_RELATION_WAS_ACCEPTED_SUCCESSFULLY,
+        HTTPStatus.OK, f"{messages.MENTORSHIP_RELATION_WAS_ACCEPTED_SUCCESSFULLY}"
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.FORBIDDEN.value,
-        "%s\n%s\n%s\n%s"
-        % (
-            messages.NOT_PENDING_STATE_RELATION,
-            messages.CANT_ACCEPT_MENTOR_REQ_SENT_BY_USER,
-            messages.CANT_ACCEPT_UNINVOLVED_MENTOR_RELATION,
-            messages.USER_IS_INVOLVED_IN_A_MENTORSHIP_RELATION,
-        ),
+        HTTPStatus.FORBIDDEN,
+        f"{messages.NOT_PENDING_STATE_RELATION}\n"
+        f"{messages.CANT_ACCEPT_MENTOR_REQ_SENT_BY_USER}\n"
+        f"{messages.CANT_ACCEPT_UNINVOLVED_MENTOR_RELATION}\n"
+        f"{messages.USER_IS_INVOLVED_IN_A_MENTORSHIP_RELATION}",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-        ),
+        HTTPStatus.UNAUTHORIZED,
+        f"{messages.TOKEN_HAS_EXPIRED}\n"
+        f"{messages.TOKEN_IS_INVALID}\n"
+        f"{messages.AUTHORISATION_TOKEN_IS_MISSING}",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND.value,
-        "%s" % messages.MENTORSHIP_RELATION_REQUEST_DOES_NOT_EXIST,
+        HTTPStatus.NOT_FOUND, f"{messages.MENTORSHIP_RELATION_REQUEST_DOES_NOT_EXIST}"
     )
     def put(cls, request_id):
         """
@@ -252,30 +227,22 @@ class RejectMentorshipRelation(Resource):
     @mentorship_relation_ns.doc("reject_mentorship_relation")
     @mentorship_relation_ns.expect(auth_header_parser)
     @mentorship_relation_ns.response(
-        HTTPStatus.OK.value,
-        "%s" % messages.MENTORSHIP_RELATION_WAS_REJECTED_SUCCESSFULLY,
+        HTTPStatus.OK, f"{messages.MENTORSHIP_RELATION_WAS_REJECTED_SUCCESSFULLY}"
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.FORBIDDEN.value,
-        "%s\n%s\n%s"
-        % (
-            messages.NOT_PENDING_STATE_RELATION,
-            messages.USER_CANT_REJECT_REQUEST_SENT_BY_USER,
-            messages.CANT_REJECT_UNINVOLVED_RELATION_REQUEST,
-        ),
+        HTTPStatus.FORBIDDEN,
+        f"{messages.NOT_PENDING_STATE_RELATION}\n"
+        f"{messages.USER_CANT_REJECT_REQUEST_SENT_BY_USER}\n"
+        f"{messages.CANT_REJECT_UNINVOLVED_RELATION_REQUEST}",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-        ),
+        HTTPStatus.UNAUTHORIZED,
+        f"{messages.TOKEN_HAS_EXPIRED}\n"
+        f"{messages.TOKEN_IS_INVALID}\n"
+        f"{messages.AUTHORISATION_TOKEN_IS_MISSING}",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND.value,
-        "%s" % messages.MENTORSHIP_RELATION_REQUEST_DOES_NOT_EXIST,
+        HTTPStatus.NOT_FOUND, f"{messages.MENTORSHIP_RELATION_REQUEST_DOES_NOT_EXIST}"
     )
     def put(cls, request_id):
         """
@@ -304,26 +271,21 @@ class CancelMentorshipRelation(Resource):
     @mentorship_relation_ns.doc("cancel_mentorship_relation")
     @mentorship_relation_ns.expect(auth_header_parser)
     @mentorship_relation_ns.response(
-        HTTPStatus.OK.value,
-        "%s" % messages.MENTORSHIP_RELATION_WAS_CANCELLED_SUCCESSFULLY,
+        HTTPStatus.OK, f"{messages.MENTORSHIP_RELATION_WAS_CANCELLED_SUCCESSFULLY}"
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.BAD_REQUEST.value,
-        "%s\n%s"
-        % (messages.UNACCEPTED_STATE_RELATION, messages.CANT_CANCEL_UNINVOLVED_REQUEST),
+        HTTPStatus.BAD_REQUEST,
+        f"{messages.UNACCEPTED_STATE_RELATION}\n"
+        f"{messages.CANT_CANCEL_UNINVOLVED_REQUEST}",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-        ),
+        HTTPStatus.UNAUTHORIZED,
+        f"{messages.TOKEN_HAS_EXPIRED}\n"
+        f"{messages.TOKEN_IS_INVALID}\n"
+        f"{messages.AUTHORISATION_TOKEN_IS_MISSING}",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND.value,
-        "%s" % messages.MENTORSHIP_RELATION_REQUEST_DOES_NOT_EXIST,
+        HTTPStatus.NOT_FOUND, f"{messages.MENTORSHIP_RELATION_REQUEST_DOES_NOT_EXIST}"
     )
     def put(cls, request_id):
         """
@@ -352,29 +314,21 @@ class DeleteMentorshipRelation(Resource):
     @mentorship_relation_ns.doc("delete_mentorship_relation")
     @mentorship_relation_ns.expect(auth_header_parser)
     @mentorship_relation_ns.response(
-        HTTPStatus.OK.value,
-        "%s" % messages.MENTORSHIP_RELATION_WAS_DELETED_SUCCESSFULLY,
+        HTTPStatus.OK, f"{messages.MENTORSHIP_RELATION_WAS_DELETED_SUCCESSFULLY}"
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.FORBIDDEN.value,
-        "%s\n%s"
-        % (
-            messages.NOT_PENDING_STATE_RELATION,
-            messages.CANT_DELETE_UNINVOLVED_REQUEST,
-        ),
+        HTTPStatus.FORBIDDEN,
+        f"{messages.NOT_PENDING_STATE_RELATION}\n"
+        f"{messages.CANT_DELETE_UNINVOLVED_REQUEST}",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-        ),
+        HTTPStatus.UNAUTHORIZED,
+        f"{messages.TOKEN_HAS_EXPIRED}\n"
+        f"{messages.TOKEN_IS_INVALID}\n"
+        f"{messages.AUTHORISATION_TOKEN_IS_MISSING}",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND.value,
-        "%s" % messages.MENTORSHIP_RELATION_REQUEST_DOES_NOT_EXIST,
+        HTTPStatus.NOT_FOUND, f"{messages.MENTORSHIP_RELATION_REQUEST_DOES_NOT_EXIST}"
     )
     def delete(cls, request_id):
         """
@@ -408,13 +362,10 @@ class ListPastMentorshipRelations(Resource):
         model=mentorship_request_response_body,
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-        ),
+        HTTPStatus.UNAUTHORIZED,
+        f"{messages.TOKEN_HAS_EXPIRED}\n"
+        f"{messages.TOKEN_IS_INVALID}\n"
+        f"{messages.AUTHORISATION_TOKEN_IS_MISSING}",
     )
     @mentorship_relation_ns.marshal_list_with(
         mentorship_request_response_body,
@@ -450,13 +401,10 @@ class ListCurrentMentorshipRelation(Resource):
         model=mentorship_request_response_body,
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-        ),
+        HTTPStatus.UNAUTHORIZED,
+        f"{messages.TOKEN_HAS_EXPIRED}\n"
+        f"{messages.TOKEN_IS_INVALID}\n"
+        f"{messages.AUTHORISATION_TOKEN_IS_MISSING}",
     )
     def get(cls):
         """
@@ -498,13 +446,10 @@ class ListPendingMentorshipRequests(Resource):
         description="Success",
     )
     @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-        ),
+        HTTPStatus.UNAUTHORIZED,
+        f"{messages.TOKEN_HAS_EXPIRED}\n"
+        f"{messages.TOKEN_IS_INVALID}\n"
+        f"{messages.AUTHORISATION_TOKEN_IS_MISSING}",
     )
     def get(cls):
         """
@@ -521,371 +466,3 @@ class ListPendingMentorshipRequests(Resource):
         response = DAO.list_pending_mentorship_relations(user_id)
 
         return response
-
-
-@mentorship_relation_ns.route("mentorship_relation/<int:request_id>/task")
-class CreateTask(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.doc("create_task_in_mentorship_relation")
-    @mentorship_relation_ns.expect(auth_header_parser, create_task_request_body)
-    @mentorship_relation_ns.response(
-        HTTPStatus.CREATED.value, "%s" % messages.TASK_WAS_CREATED_SUCCESSFULLY
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.FORBIDDEN.value, "%s" % messages.UNACCEPTED_STATE_RELATION
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-        ),
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.FORBIDDEN.value,
-        "%s" % messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION,
-    )
-    def post(cls, request_id):
-        """
-        Create a task for a mentorship relation.
-
-        Input:
-        1. Header: valid access token
-        2. Path: ID of request for which task is being created (request_id)
-        3. Body: JSON object containing description of task.
-
-        Returns:
-        Success or failure message. It gets added to GET /tasks endpoint and
-        is visible to the other person in the mentorship relation.
-        """
-
-        # TODO check if user id is well parsed, if it is an integer
-
-        user_id = get_jwt_identity()
-        request_body = request.json
-
-        is_valid = CreateTask.is_valid_data(request_body)
-
-        if is_valid != {}:
-            return is_valid, HTTPStatus.BAD_REQUEST
-
-        response = TaskDAO.create_task(
-            user_id=user_id, mentorship_relation_id=request_id, data=request_body
-        )
-
-        return response
-
-    @staticmethod
-    def is_valid_data(data):
-
-        if "description" not in data:
-            return messages.DESCRIPTION_FIELD_IS_MISSING
-
-        return {}
-
-
-@mentorship_relation_ns.route("mentorship_relation/<int:request_id>/task/<int:task_id>")
-class DeleteTask(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.doc("delete_task_in_mentorship_relation")
-    @mentorship_relation_ns.expect(auth_header_parser)
-    @mentorship_relation_ns.response(
-        HTTPStatus.OK.value, "%s" % messages.TASK_WAS_DELETED_SUCCESSFULLY
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-            messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION,
-        ),
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND.value,
-        "%s\n%s"
-        % (messages.MENTORSHIP_RELATION_DOES_NOT_EXIST, messages.TASK_DOES_NOT_EXIST),
-    )
-    def delete(cls, request_id, task_id):
-        """
-        Delete a task.
-
-        Input:
-        1. Header: valid access token
-        2. Path: ID of the task to be deleted (task_id) and it ID of the associated
-        mentorship relation (request_id).
-        3. Body: JSON object containing description of task.
-
-        Returns:
-        Success or failure message. Task is deleted if request is successful.
-        """
-
-        # TODO check if user id is well parsed, if it is an integer
-
-        user_id = get_jwt_identity()
-
-        response = TaskDAO.delete_task(
-            user_id=user_id, mentorship_relation_id=request_id, task_id=task_id
-        )
-
-        return response
-
-
-@mentorship_relation_ns.route("mentorship_relation/<int:request_id>/tasks")
-class ListTasks(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.doc("list_tasks_in_mentorship_relation")
-    @mentorship_relation_ns.expect(auth_header_parser)
-    @mentorship_relation_ns.response(
-        HTTPStatus.OK.value,
-        "List tasks from a mentorship relation with success.",
-        model=list_tasks_response_body,
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-            messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION,
-        ),
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND.value, "%s" % messages.MENTORSHIP_RELATION_DOES_NOT_EXIST
-    )
-    def get(cls, request_id):
-        """
-        List all tasks from a mentorship relation.
-
-        Input:
-        1. Header: valid access token
-        2. Path: ID of the mentorship relation for which tasks are to be
-        displayed(request_id). The user must be involved in this relation.
-
-        Returns:
-        JSON array containing task details as objects is displayed on success.
-        """
-
-        # TODO check if user id is well parsed, if it is an integer
-
-        user_id = get_jwt_identity()
-
-        response = TaskDAO.list_tasks(
-            user_id=user_id, mentorship_relation_id=request_id
-        )
-
-        if isinstance(response, tuple):
-            return response
-
-        return marshal(response, list_tasks_response_body), HTTPStatus.OK
-
-
-@mentorship_relation_ns.route(
-    "mentorship_relation/<int:request_id>/task/<int:task_id>/complete"
-)
-class UpdateTask(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.doc("update_task_in_mentorship_relation")
-    @mentorship_relation_ns.expect(auth_header_parser)
-    @mentorship_relation_ns.response(
-        HTTPStatus.OK.value, "%s" % messages.TASK_WAS_ACHIEVED_SUCCESSFULLY
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.BAD_REQUEST.value, "%s" % messages.TASK_WAS_ALREADY_ACHIEVED
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED.value,
-        "%s\n%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-            messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION,
-        ),
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND.value,
-        "%s\n%s"
-        % (messages.MENTORSHIP_RELATION_DOES_NOT_EXIST, messages.TASK_DOES_NOT_EXIST),
-    )
-    def put(cls, request_id, task_id):
-        """
-        Update a task to mark it as complate
-
-        Input:
-        1. Header: valid access token
-        2. Path: ID of task (task_id) and ID of the associated mentorship
-        relation (request_id). The user must be involved in this relation.
-        3. Body:
-
-        Returns:
-        Success or failure message. The task is marked as complete if succesful.
-        """
-
-        # TODO check if user id is well parsed, if it is an integer
-
-        user_id = get_jwt_identity()
-
-        response = TaskDAO.complete_task(
-            user_id=user_id, mentorship_relation_id=request_id, task_id=task_id
-        )
-
-        return response
-
-
-@mentorship_relation_ns.route(
-    "mentorship_relation/<int:relation_id>/task/<int:task_id>/comment"
-)
-class CreateTaskComment(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.expect(auth_header_parser, task_comment_model)
-    @mentorship_relation_ns.doc(
-        responses={
-            HTTPStatus.CREATED.value: f"{messages.TASK_COMMENT_WAS_CREATED_SUCCESSFULLY}",
-            HTTPStatus.BAD_REQUEST.value: f"{messages.COMMENT_FIELD_IS_MISSING}<br>"
-            f"{messages.COMMENT_NOT_IN_STRING_FORMAT}<br>"
-            f"{ {'message': get_length_validation_error_message('comment', None, COMMENT_MAX_LENGTH)}}<br>"
-            f"{messages.UNACCEPTED_STATE_RELATION}",
-            HTTPStatus.UNAUTHORIZED.value: f"{messages.TOKEN_HAS_EXPIRED}<br>"
-            f"{messages.TOKEN_IS_INVALID}<br>"
-            f"{messages.AUTHORISATION_TOKEN_IS_MISSING}<br>"
-            f"{messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION}",
-            HTTPStatus.NOT_FOUND.value: f"{messages.USER_DOES_NOT_EXIST}<br>"
-            f"{messages.MENTORSHIP_RELATION_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_DOES_NOT_EXIST}",
-        }
-    )
-    def post(cls, relation_id, task_id):
-        """
-        Creates a new task comment.
-        """
-
-        data = request.json
-
-        is_valid = validate_task_comment_request_data(data)
-        if is_valid != {}:
-            return is_valid, HTTPStatus.BAD_REQUEST
-
-        comment = data["comment"]
-        return TaskCommentDAO.create_task_comment(
-            get_jwt_identity(), task_id, relation_id, comment
-        )
-
-
-@mentorship_relation_ns.route(
-    "mentorship_relation/<int:relation_id>/task/<int:task_id>/comment/<int:comment_id>"
-)
-class TaskComment(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.expect(auth_header_parser, task_comment_model)
-    @mentorship_relation_ns.doc(
-        responses={
-            HTTPStatus.OK.value: f"{messages.TASK_COMMENT_WAS_UPDATED_SUCCESSFULLY}",
-            HTTPStatus.BAD_REQUEST.value: f"{messages.COMMENT_FIELD_IS_MISSING}<br>"
-            f"{messages.COMMENT_NOT_IN_STRING_FORMAT}<br>"
-            f"{ {'message':get_length_validation_error_message('comment', None, COMMENT_MAX_LENGTH)}}<br>"
-            f"{messages.UNACCEPTED_STATE_RELATION}<br>"
-            f"{messages.TASK_COMMENT_WAS_NOT_CREATED_BY_YOU}",
-            HTTPStatus.UNAUTHORIZED.value: f"{messages.TOKEN_HAS_EXPIRED}<br>"
-            f"{messages.TOKEN_IS_INVALID}<br>"
-            f"{messages.AUTHORISATION_TOKEN_IS_MISSING}<br>"
-            f"{messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION}",
-            HTTPStatus.NOT_FOUND.value: f"{messages.USER_DOES_NOT_EXIST}<br>"
-            f"{messages.MENTORSHIP_RELATION_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_COMMENT_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_COMMENT_WITH_GIVEN_TASK_ID_DOES_NOT_EXIST}",
-        }
-    )
-    def put(cls, relation_id, task_id, comment_id):
-        """
-        Modifies the task comment.
-        """
-
-        data = request.json
-
-        is_valid = validate_task_comment_request_data(data)
-        if is_valid != {}:
-            return is_valid, HTTPStatus.BAD_REQUEST
-
-        comment = data["comment"]
-        return TaskCommentDAO.modify_comment(
-            get_jwt_identity(), comment_id, task_id, relation_id, comment
-        )
-
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.expect(auth_header_parser)
-    @mentorship_relation_ns.doc(
-        responses={
-            HTTPStatus.OK.value: f"{messages.TASK_COMMENT_WAS_DELETED_SUCCESSFULLY}",
-            HTTPStatus.BAD_REQUEST.value: f"{messages.UNACCEPTED_STATE_RELATION}",
-            HTTPStatus.UNAUTHORIZED.value: f"{messages.TOKEN_HAS_EXPIRED}<br>"
-            f"{messages.TOKEN_IS_INVALID}<br>"
-            f"{messages.AUTHORISATION_TOKEN_IS_MISSING}<br>"
-            f"{messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION}",
-            HTTPStatus.FORBIDDEN.value: f"{messages.TASK_COMMENT_WAS_NOT_CREATED_BY_YOU_DELETE}",
-            HTTPStatus.NOT_FOUND.value: f"{messages.USER_DOES_NOT_EXIST}<br>"
-            f"{messages.MENTORSHIP_RELATION_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_COMMENT_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_COMMENT_WITH_GIVEN_TASK_ID_DOES_NOT_EXIST}",
-        }
-    )
-    def delete(cls, relation_id, task_id, comment_id):
-        """
-        Deletes the task comment.
-        """
-
-        return TaskCommentDAO.delete_comment(
-            get_jwt_identity(), comment_id, task_id, relation_id
-        )
-
-
-@mentorship_relation_ns.route(
-    "mentorship_relation/<int:relation_id>/task/<int:task_id>/comments/"
-)
-class TaskComments(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.expect(auth_header_parser)
-    @mentorship_relation_ns.response(
-        HTTPStatus.OK.value,
-        f"{ messages.LIST_TASK_COMMENTS_WITH_SUCCESS}",
-        task_comments_model,
-    )
-    @mentorship_relation_ns.doc(
-        responses={
-            HTTPStatus.BAD_REQUEST.value: f"{messages.UNACCEPTED_STATE_RELATION}",
-            HTTPStatus.UNAUTHORIZED.value: f"{messages.TOKEN_HAS_EXPIRED}<br>"
-            f"{messages.TOKEN_IS_INVALID}<br>"
-            f"{messages.AUTHORISATION_TOKEN_IS_MISSING}<br>"
-            f"{messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION}",
-            HTTPStatus.NOT_FOUND.value: f"{messages.USER_DOES_NOT_EXIST}<br>"
-            f"{messages.MENTORSHIP_RELATION_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_DOES_NOT_EXIST}",
-        }
-    )
-    def get(cls, relation_id, task_id):
-        """
-        Lists the task comments.
-        """
-
-        response = TaskCommentDAO.get_all_task_comments_by_task_id(
-            get_jwt_identity(), task_id, relation_id
-        )
-
-        if isinstance(response, tuple):
-            return response
-        else:
-            return marshal(response, task_comments_model), HTTPStatus.OK
